[TASK] Align namespace of console input validators
