+++ src/Command/InitConfigCommand.php
@@ -25,10 +25,11 @@ namespace CPSIT\FrontendAssetHandler\Command;
 
 use CPSIT\FrontendAssetHandler\Asset;
 use CPSIT\FrontendAssetHandler\Config;
+use CPSIT\FrontendAssetHandler\Console;
 use CPSIT\FrontendAssetHandler\Exception;
 use CPSIT\FrontendAssetHandler\Helper;
 use CPSIT\FrontendAssetHandler\Json;
-use Symfony\Component\Console;
+use Symfony\Component\Console as SymfonyConsole;
 use Symfony\Component\DependencyInjection;
 
 use function array_key_last;
@@ -44,11 +45,11 @@ use function strtolower;
  * @author Elias Häußler <e.haeussler@familie-redlich.de>
  * @license GPL-3.0-or-later
  */
-final class InitConfigCommand extends Console\Command\Command
+final class InitConfigCommand extends SymfonyConsole\Command\Command
 {
     private const SUCCESSFUL = 0;
 
-    private Console\Style\SymfonyStyle $io;
+    private SymfonyConsole\Style\SymfonyStyle $io;
 
     public function __construct(
         private readonly Config\ConfigFacade $configFacade,
@@ -69,91 +70,91 @@ final class InitConfigCommand extends Console\Command\Command
         $this->addOption(
             'source-type',
             null,
-            Console\Input\InputOption::VALUE_REQUIRED,
+            SymfonyConsole\Input\InputOption::VALUE_REQUIRED,
             'Type of the asset source, resolves to a supported asset provider',
         );
         $this->addOption(
             'source-url',
             null,
-            Console\Input\InputOption::VALUE_REQUIRED,
+            SymfonyConsole\Input\InputOption::VALUE_REQUIRED,
             'URL to locate the asset source files, can contain placeholders in the form {<config key>}',
         );
         $this->addOption(
             'source-revision-url',
             null,
-            Console\Input\InputOption::VALUE_REQUIRED,
+            SymfonyConsole\Input\InputOption::VALUE_REQUIRED,
             'URL to locate the revision of asset source files, can contain placeholders in the form {<config key>}',
         );
         $this->addOption(
             'source-config-extra',
             null,
-            Console\Input\InputOption::VALUE_REQUIRED,
+            SymfonyConsole\Input\InputOption::VALUE_REQUIRED,
             'Additional configuration for the asset source definition, should be a JSON-encoded string',
         );
 
         $this->addOption(
             'target-type',
             null,
-            Console\Input\InputOption::VALUE_REQUIRED,
+            SymfonyConsole\Input\InputOption::VALUE_REQUIRED,
             'Type of the asset target, resolves to a supported asset processor',
         );
         $this->addOption(
             'target-path',
             null,
-            Console\Input\InputOption::VALUE_REQUIRED,
+            SymfonyConsole\Input\InputOption::VALUE_REQUIRED,
             'Path where to extract fetched assets, can be either absolute or relative to the config file',
         );
         $this->addOption(
             'target-revision-file',
             null,
-            Console\Input\InputOption::VALUE_REQUIRED,
+            SymfonyConsole\Input\InputOption::VALUE_REQUIRED,
             'Filename of the asset target\'s revision file',
             Asset\Definition\Target::DEFAULT_REVISION_FILE,
         );
         $this->addOption(
             'target-config-extra',
             null,
-            Console\Input\InputOption::VALUE_REQUIRED,
+            SymfonyConsole\Input\InputOption::VALUE_REQUIRED,
             'Additional configuration for the asset target definition, should be a JSON-encoded string',
         );
 
         $this->addOption(
             'handler-type',
             null,
-            Console\Input\InputOption::VALUE_REQUIRED,
+            SymfonyConsole\Input\InputOption::VALUE_REQUIRED,
             'Type of the asset handler, resolves to a supported asset handler',
         );
 
         $this->addOption(
             'vcs-type',
             null,
-            Console\Input\InputOption::VALUE_REQUIRED,
+            SymfonyConsole\Input\InputOption::VALUE_REQUIRED,
             'Type of the asset\'s VCS, resolves to a supported VCS provider',
         );
         $this->addOption(
             'vcs-config-extra',
             null,
-            Console\Input\InputOption::VALUE_REQUIRED,
+            SymfonyConsole\Input\InputOption::VALUE_REQUIRED,
             'Additional configuration for the asset VCS definition, should be a JSON-encoded string',
         );
 
         $this->addOption(
             'definition-id',
             null,
-            Console\Input\InputOption::VALUE_REQUIRED,
+            SymfonyConsole\Input\InputOption::VALUE_REQUIRED,
             'ID of the asset definition to be added to the asset configuration file',
             0,
         );
     }
 
-    protected function initialize(Console\Input\InputInterface $input, Console\Output\OutputInterface $output): void
+    protected function initialize(SymfonyConsole\Input\InputInterface $input, SymfonyConsole\Output\OutputInterface $output): void
     {
-        $this->io = new Console\Style\SymfonyStyle($input, $output);
+        $this->io = new SymfonyConsole\Style\SymfonyStyle($input, $output);
     }
 
-    protected function interact(Console\Input\InputInterface $input, Console\Output\OutputInterface $output): void
+    protected function interact(SymfonyConsole\Input\InputInterface $input, SymfonyConsole\Output\OutputInterface $output): void
     {
-        /** @var Console\Helper\QuestionHelper $helper */
+        /** @var SymfonyConsole\Helper\QuestionHelper $helper */
         $helper = $this->getHelper('question');
 
         $output->writeln([
@@ -174,15 +175,15 @@ final class InitConfigCommand extends Console\Command\Command
         $input->setOption('source-type', $helper->ask($input, $output, $question));
 
         $question = $this->createQuestion('URL', $input->getOption('source-url'));
-        $question->setValidator(Validators\UrlValidator::validate(...));
+        $question->setValidator(Console\Input\Validator\UrlValidator::validate(...));
         $input->setOption('source-url', $helper->ask($input, $output, $question));
 
         $question = $this->createQuestion('Revision URL', $input->getOption('source-revision-url'));
-        $question->setValidator(Validators\UrlValidator::validate(...));
+        $question->setValidator(Console\Input\Validator\UrlValidator::validate(...));
         $input->setOption('source-revision-url', $helper->ask($input, $output, $question));
 
         $question = $this->createQuestion('Additional config', $input->getOption('source-config-extra'));
-        $question->setValidator(Validators\JsonValidator::validate(...));
+        $question->setValidator(Console\Input\Validator\JsonValidator::validate(...));
         $input->setOption('source-config-extra', $helper->ask($input, $output, $question));
 
         $this->io->newLine();
@@ -196,15 +197,15 @@ final class InitConfigCommand extends Console\Command\Command
         $input->setOption('target-type', $helper->ask($input, $output, $question));
 
         $question = $this->createQuestion('Path', $input->getOption('target-path'));
-        $question->setValidator(Validators\NotEmptyValidator::validate(...));
+        $question->setValidator(Console\Input\Validator\NotEmptyValidator::validate(...));
         $input->setOption('target-path', $helper->ask($input, $output, $question));
 
         $question = $this->createQuestion('Revision file', $input->getOption('target-revision-file'));
-        $question->setValidator(Validators\NotEmptyValidator::validate(...));
+        $question->setValidator(Console\Input\Validator\NotEmptyValidator::validate(...));
         $input->setOption('target-revision-file', $helper->ask($input, $output, $question));
 
         $question = $this->createQuestion('Additional config', $input->getOption('target-config-extra'));
-        $question->setValidator(Validators\JsonValidator::validate(...));
+        $question->setValidator(Console\Input\Validator\JsonValidator::validate(...));
         $input->setOption('target-config-extra', $helper->ask($input, $output, $question));
 
         $this->io->newLine();
@@ -240,14 +241,14 @@ final class InitConfigCommand extends Console\Command\Command
         $input->setOption('vcs-type', $helper->ask($input, $output, $question));
 
         $question = $this->createQuestion('Additional config', $input->getOption('vcs-config-extra'));
-        $question->setValidator(Validators\JsonValidator::validate(...));
+        $question->setValidator(Console\Input\Validator\JsonValidator::validate(...));
         $input->setOption('vcs-config-extra', $helper->ask($input, $output, $question));
     }
 
-    protected function execute(Console\Input\InputInterface $input, Console\Output\OutputInterface $output): int
+    protected function execute(SymfonyConsole\Input\InputInterface $input, SymfonyConsole\Output\OutputInterface $output): int
     {
         if (!$input->isInteractive()) {
-            throw new Console\Exception\RuntimeException('This console command can only be run in interactive mode.', 1662395429);
+            throw new SymfonyConsole\Exception\RuntimeException('This console command can only be run in interactive mode.', 1662395429);
         }
 
         // Resolve config file
@@ -321,10 +322,10 @@ final class InitConfigCommand extends Console\Command\Command
     }
 
     private function selectConfigFileAndDefinitionId(
-        Console\Input\InputInterface $input,
-        Console\Output\OutputInterface $output,
+        SymfonyConsole\Input\InputInterface $input,
+        SymfonyConsole\Output\OutputInterface $output,
     ): void {
-        /** @var Console\Helper\QuestionHelper $helper */
+        /** @var SymfonyConsole\Helper\QuestionHelper $helper */
         $helper = $this->getHelper('question');
 
         // Load config file
@@ -373,7 +374,7 @@ final class InitConfigCommand extends Console\Command\Command
             return;
         }
 
-        $question = new Console\Question\Question('<info>Path to the new config file</info>: ');
+        $question = new SymfonyConsole\Question\Question('<info>Path to the new config file</info>: ');
 
         $input->setOption('config', $helper->ask($input, $output, $question));
         $input->setOption('definition-id', 0);
@@ -397,21 +398,21 @@ final class InitConfigCommand extends Console\Command\Command
         return $this->configParser->parse($instructions);
     }
 
-    private function createQuestion(string $label, mixed $default = null, string $alternative = null): Console\Question\Question
+    private function createQuestion(string $label, mixed $default = null, string $alternative = null): SymfonyConsole\Question\Question
     {
         $label = $this->decorateQuestionLabel($label, $default, $alternative);
 
-        return new Console\Question\Question($label, $default);
+        return new SymfonyConsole\Question\Question($label, $default);
     }
 
     /**
      * @param array<string, string> $choices
      */
-    private function createChoiceQuestion(string $label, array $choices, mixed $default = null): Console\Question\ChoiceQuestion
+    private function createChoiceQuestion(string $label, array $choices, mixed $default = null): SymfonyConsole\Question\ChoiceQuestion
     {
         $label = $this->decorateQuestionLabel($label, $default);
 
-        return new Console\Question\ChoiceQuestion($label, $choices, $default);
+        return new SymfonyConsole\Question\ChoiceQuestion($label, $choices, $default);
     }
 
     private function decorateQuestionLabel(string $label, mixed $default, string $alternative = null): string
