+++ src/Console/Input/Validator/UrlValidator.php
@@ -21,15 +21,29 @@ declare(strict_types=1);
  * along with this program. If not, see <https://www.gnu.org/licenses/>.
  */
 
-namespace CPSIT\FrontendAssetHandler\Command\Validators;
+namespace CPSIT\FrontendAssetHandler\Console\Input\Validator;
+
+use Webmozart\Assert;
+
+use function filter_var;
 
 /**
- * ValidatorInterface.
+ * UrlValidator.
  *
  * @author Elias Häußler <e.haeussler@familie-redlich.de>
  * @license GPL-3.0-or-later
  */
-interface ValidatorInterface
+final class UrlValidator implements ValidatorInterface
 {
-    public static function validate(mixed $value): mixed;
+    public static function validate(mixed $value): string
+    {
+        Assert\Assert::stringNotEmpty($value);
+
+        // Allow placeholders in URLs. Those will be replaced later by string interpolation.
+        $normalizedUrl = preg_replace('/{[^}]+}/', 'placeholder', $value);
+
+        Assert\Assert::notFalse(filter_var($normalizedUrl, FILTER_VALIDATE_URL), 'The given URL is invalid.');
+
+        return $value;
+    }
 }
